[change][schema] Move Item-related types into schema/item sub-package
